feat(renderer): optimize LibraryListView with content-visibility and memoization
- Extract `LibraryListItem` component from `LibraryListView`
- Implement `content-visibility: auto` and `contain-intrinsic-size` in `LibraryListItem`
- Wrap `LibraryListItem` in `React.memo` to prevent unnecessary re-renders
- Move formatter functions to `renderer/src/utils/formatters.ts` for reuse
- Add type safety to `LibraryListItem` props
